updated CSS to get rid of padding on search button
This is to be aligned with other elements. this css is restrictive to
storage-modal class only

bootstraps row class gives a margin to the right

This causes the storage to always have a horizontal scroll bar.  this fixes the problem when using storage-modal on the body tag.

added iframe.modal-dialog selector

This is to correct the css on the new modal changes in style-guide.

updated dist to reflect changes